Adding in 'Task' structure, noticed performance issues
Working in the 'task' structure, which will serve as a way to represent the overall goal of an entity, and will decide what actions it will take. this is the real 'AI' part of the code.

Got rid of some leftover code that was causing issues.

Noticed huge performance drops when you push the game i.e hold any action. It is not consistent, could be due to poorly placed thread.sleep() calls but im pretty sure it's just the raytracing being stupid expensive. Need to take another look at that, there are some huge optimizations gthat could be made