AI POC With player visuals works. Fixed A* problem by decreasing weight of heuristic.
Was having a problem where the AI would priorize optimising heuristic early on, leading to a path that was one or two blocks directly towards the target, even if that led into a wall - it would then continue towards an opening and get to target. This no longer happens, just divided heuristic by 10.