Added update tracking, structure for a complex AI framework and improved 'Explore' AI.
Major changes and refactors.
Creature memory now tracks whether it has visited cells since they have changed (for now 'changed' refers to if they have recently become enterable)
This means that the Explore AI can re investigate areas if a door has been open, explored, closed, and opened again.

The AI framework consists of a list of conditional branches  (A predicate 'query' on the creature, attached to a supplier for a new State) sorted by priority. These will be iterated across every update, retrieving what the highest priority and VALID 'State' is. The Validity/Availability to move to a new state depends on a Predicate attached to each branch (i.e per-entity), whereas the ability to continue a current state (if it is the highest priority state available) is checked with the 'canContinue' (per-state).

AI States have an optional inner class they can extend to export information of the state.

It's important to note that if the current state is ok to proceed, a new instance of the state is not created! When a state is overwritten by another, it should be discarded