Fixed perpendicular wall rendering bug
Previously walls beyond the FOV, perpendicular and adjacent to walls that were in FOV were rendering. These have been removed with a check that removes the old 'delta' that was added for raytracing to be able to hop to the next cell.

There are still walls rendering that are parallel to walls that are within FOV, working on that now.