Fixed floating point errors in tile intersection algorithm
When intersections were getting really close to integer points the distances from one intersection to the next was getting so small that it resulted in floating point errors, causing the next intersection point to be improperly calculated, sometimes repeating in unending cycle, adding the same tiles back and forth.

Added a little bump, where if the distance to intersection is too small, a bump in the dy/dx direction is added, causing the ray to be 'bumped' a little of track, small enough to have no impact on rest of the line. If worried about this stacking (if 100s of blocks in distance...), just add an accumulator for this and subtract from it later down the ray.