Framework complete, with basic datastreams for Cells and Entities skeleton.
Took a lot of thinking and redesigning but I'm happy with this.
For each Datastream you need to define an associated DataParameter and Data CLASS. be careful about the generic bounds.
1 Instance of each datastream as well. Currently statically typed, wondering if I can write a fancy map for this...
Each entity has 1 or more sensors.
Main game loop should query Creatures for sensors. Sensors are 'polled', and those in turn retrieve a DataParameter obj from Entity.
Sensor then sends DataParameter obj to associated Datastreamm which then returns a Data object. This is then handed to the Brain, which should have strategies to deal with different Data types.