Fixed/identified remaining flicker issues
The client side decoder is too slow. This is ok though because it's a stupid design and I'm going to redo it completely (we shouldn't be serializing server internals anyways...) will come up with a lightweight set of packets that can be sent to client that will offload a ton of complexity from networking/client side.